+++ mobile/src/navigation/types.tsx
@@ -1,22 +1,7 @@
 import { StackScreenProps } from '@react-navigation/stack';
-import { Camera, LatLng } from 'react-native-maps';
-import { RouterOutput } from '@src/utils/trpc';
 
-export type RootStackParamList = {
-  Home: undefined;
-  Search: undefined;
-  MapType: undefined;
-  Details: {
-    highline: RouterOutput['highline']['createHighline'];
-  };
-  LocationPicker: {
-    camera?: Camera;
-  };
-  HighlineFormScreen: {
-    lenght: string;
-    markers: LatLng[];
-  };
-};
+import { RootStackParamList } from './app.routes';
+import { AuthStackParamList } from './auth.routes';
 
 export type HomeScreenProps = StackScreenProps<RootStackParamList, 'Home'>;
 
@@ -29,3 +14,7 @@ export type DetailScreenProps = StackScreenProps<RootStackParamList, 'Details'>;
 export type LocationPickerScreenProps = StackScreenProps<RootStackParamList, 'LocationPicker'>;
 
 export type HighlineFormScreenProps = StackScreenProps<RootStackParamList, 'HighlineFormScreen'>;
+
+export type SignInScreenProps = StackScreenProps<AuthStackParamList, 'SignIn'>;
+
+export type SignUpScreenProps = StackScreenProps<AuthStackParamList, 'SignUp'>;
